refactor: high level mod structure
