--- packages/ic-response-verification/src/cbor.rs
@@ -1,5 +1,0 @@
-pub mod certificate;
-pub mod hash_tree;
-
-mod cbor;
-pub(crate) use cbor::*;
