+++ packages/ic-response-verification/src/types/mod.rs
@@ -1,21 +1,21 @@
 //! Public types used for response verification.
 
 /// Types to represent parsed CEL expressions.
-pub mod certification;
+mod certification;
 pub use certification::*;
 
 /// Types to represent response objects used for certification.
-pub mod request;
+mod request;
 pub use request::*;
 
 /// Types to represent request objects used for certification.
-pub mod response;
+mod response;
 pub use response::*;
 
 /// Types to represent the result of verifying a request/response pair's certification.
-pub mod certification_result;
+mod certification_result;
 pub use certification_result::*;
 
 /// Types to represent a certified response that clients can use to determine which parts of a response are safe to use.
-pub mod certified_response;
+mod certified_response;
 pub use certified_response::*;
