+++ packages/ic-response-verification/src/verification/mod.rs
@@ -1,0 +1,8 @@
+//! The primary entry point for the repsonse verification API.
+
+mod body;
+mod certificate_header;
+mod certificate_header_field;
+
+mod verify_request_response_pair;
+pub use verify_request_response_pair::*;
